Create 78_Subsets.py
78.

To solve the subsets problem, you can use backtracking where for each element in the array, you make two choices: either include it in the current subset or skip it. For each valid combination you build during this process, add it to your result list, being sure to make a copy of the current subset since it will be modified during backtracking.